updated delete todo item command to check for ownership

+++ todo-app-backend/src/Application/TodoItems/Commands/DeleteTodoItem/DeleteTodoItem.cs
@@ -1,4 +1,5 @@
 ﻿using todo_app_backend.Application.Common.Interfaces;
+using todo_app_backend.Domain.Entities;
 
 namespace todo_app_backend.Application.TodoItems.Commands.DeleteTodoItem;
 
@@ -7,18 +8,26 @@ public record DeleteTodoItemCommand(int Id) : IRequest;
 public class DeleteTodoItemCommandHandler : IRequestHandler<DeleteTodoItemCommand>
 {
     private readonly IApplicationDbContext _context;
+    private readonly IUser _user;
 
-    public DeleteTodoItemCommandHandler(IApplicationDbContext context)
+    public DeleteTodoItemCommandHandler(IApplicationDbContext context, IUser user)
     {
         _context = context;
+        _user = user;
     }
 
     public async Task Handle(DeleteTodoItemCommand request, CancellationToken cancellationToken)
     {
         var entity = await _context.TodoItems
-            .FindAsync([request.Id], cancellationToken);
+            .Include(i => i.List)
+            .FirstOrDefaultAsync(
+                i => i.Id == request.Id && i.List.CreatedBy == _user.Id,
+                cancellationToken);
 
-        Guard.Against.NotFound(request.Id, entity);
+        if (entity is null)
+        {
+            throw new NotFoundException(nameof(TodoItem), request.Id.ToString());
+        }
 
         _context.TodoItems.Remove(entity);
